--- docker/Dockerfile
@@ -1,8 +1,0 @@
-FROM google/golang:1.4
-
-MAINTAINER Derek Collison <derek@apcera.com>
-
-RUN CGO_ENABLED=0 go get -a -ldflags '-s' --installsuffix cgo github.com/apcera/gnatsd
-COPY Dockerfile.final /gopath/bin/Dockerfile
-
-CMD docker build -t apcera/gnatsd /gopath/bin
