--- docker/README.md
@@ -1,8 +1,0 @@
-# [Dockerized] (http://www.docker.com) [nats](https://registry.hub.docker.com/_/nats/)
-
-This is deprecated, please use [nats-docker](https://github.com/nats-io/nats-docker)
-
-A docker image for gnatsd. This is created as a single static executable, so there are multiple passes through Docker to first build the static executable and then to package it under an empty (scratch) base image.
-
-
-
