--- docker/linux_dev/build.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-docker build -t nats_dev .
-
